+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Alle nennenswerten Änderungen an OttoExtra. Format angelehnt an
 [Keep a Changelog](https://keepachangelog.com/de/1.1.0/); Versionierung nach
 [SemVer](https://semver.org/lang/de/).
 
+## [Unreleased] – 0.1.10
+
 ## [0.1.9] – 2026-06-30
 
 ### Hinzugefügt